Fix crash when a project has only one member cc:@evilaliv3
As the logger send e-mails to all project members BUT the
user who is performing the action, this causes an empty
recipients list that manifests as an

ArgumentError (An SMTP To address is required to send a message. Set the message smtp_envelope_to, to, cc, or bcc address.)

If the recipients list is empty, do not bother generating
an e-mail.

Refactored also the invocation, encapsulating all mail
sending inside the mailer class, rather than returning
the mail to the outside.